requirements.txt
Cheetah==2.4.4
Django==1.8.13
Landscape-Client==14.12
PAM==0.4.2
PyYAML==3.10
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
click==6.6
cloud-init==0.7.5
colorama==0.2.5
configobj==4.7.2
decorator==4.0.10
django-admin-bootstrapped==2.5.7
django-bootstrap3==7.0.1
et-xmlfile==1.0.1
geocoder==1.15.1
html5lib==0.999
jdcal==1.2
jsonpatch==1.3
jsonpointer==1.0
oauth==1.0.1
openpyxl==2.3.5
prettytable==0.7.2
psycopg2==2.4.5
pyOpenSSL==0.13
pycurl==7.19.3
pyserial==2.6
python-apt==0.9.3.5ubuntu2
python-debian==0.1.21-nmu2ubuntu2
ratelim==0.1.6
requests==2.2.1
six==1.5.2
ssh-import-id==3.21
urllib3==1.7.1
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.0.5
